+++ controllers/routers/site.js
@@ -1,6 +1,6 @@
 const { _M_ } = require('../modules/');
 const { _U_ } = require('../utilities/');
-const { ConfigsModel, MediaModel, PostsModel } = require('../../models/');
+const ModelIndex = require('../../models/');
 
 
 // controllers
@@ -13,7 +13,7 @@ site.configs = {
     return next();
   },
   PATCH: async function site_configs_PATCH(req, res) {
-    await ConfigsModel.updateConfigs(req.body.configs);                                                                 // tofix: pickup updated variables to avoid injections
+    await ModelIndex.ConfigsModel.updateConfigs(req.body.configs);                                                      // tofix: pickup updated variables to avoid injections
     return res.redirect('back');
   },
 };
@@ -24,10 +24,11 @@ site.upload = {                                                                 
     return next();
   },
   POST: [_M_.parseMultipart({ fileSize: 25 * 1048576 }), async function site_upload_POST(req, res) {
-    if (req.body.busboySlip.mes.length > 0) req.body.busboySlip.mes.forEach(mes => req.flash('error', mes));
+    if (req.body.busboySlip.mes.length > 0) req.body.busboySlip.mes.map(mes => req.flash('error', mes));
     if (req.body.busboySlip.raw.length > 0) {
-      req.body.busboySlip.raw.forEach(medium => { medium.author = req.session.user; });
-      const docs = await MediaModel.create(req.body.busboySlip.raw);
+      const slip = req.body.busboySlip.raw;
+      for (let i = slip.length - 1, media = slip[i]; i > -1; media = slip[i -= 1]) media.author = req.session.user;
+      const docs = await ModelIndex.MediaModel.create(req.body.busboySlip.raw);
       if (docs.length > 0) req.flash('info', `${docs.length} file(s) successfully uploaded!`);
     }
     return res.redirect('back');
@@ -37,18 +38,23 @@ site.upload = {                                                                 
 
 site.stack = {
   GET: async function site_stack_GET(req, res, next) {
-    const collection = req.params.stackType.toLowerCase();
+    const collection = req.params.collection.toLowerCase();
     if (!['posts', 'media'].includes(collection)) throw new _U_.error.HttpException(404);
+    //
     return _U_.express.wrapMiddleware([
       _M_.modifyHTMLTitleTag(collection),
       _M_.paginatedQuery(collection, { num: 10 }),
     ])(req, res, next);
   },
   PATCH: async function site_stack_PATCH(req, res) {
-    const $update = { $set: {} };
-    if (req.body.action === 'restored') $update.$set = { 'state.recycled': false };
-    else $update.$set = { [`state.${req.body.action}`]: true };
-    if (req.body.action) await PostsModel.update({ _id: { $in: req.body.target } }, $update, { multi: true });
+    const collection = req.params.collection.toLowerCase();
+    if (!['posts', 'media'].includes(collection)) throw new _U_.error.HttpException(404);
+    //
+    const { action, target = [] } = req.body;
+    if (action && target.length > 0) {
+      const model = ModelIndex[`${_U_.string.toCapitalized(collection)}Model`];
+      await model.update({ _id: { $in: target } }, { $set: { [`state.${action}`]: true } }, { multi: true });
+    }
     return res.redirect('back');
   },
 };